+++ pcdsdaq/daq.py
@@ -11,7 +11,6 @@ import warnings
 from importlib import import_module
 
 from ophyd.status import Status, wait as status_wait
-from ophyd.flyers import FlyerInterface
 
 logger = logging.getLogger(__name__)
 pydaq = None
@@ -41,25 +40,22 @@ def check_connect(f):
     return wrapper
 
 
-class Daq(FlyerInterface):
+class Daq:
     """
-    The LCLS1 daq as a ``Flyer`` object.
+    The LCLS1 daq as a ``bluesky``-compatible object.
 
     This uses the ``pydaq`` module to connect with a running daq instance,
     controlling it via socket commands.
+
+    It can be used as a ``Reader`` in a ``bluesky`` plan to take data at
+    discrete scan points.
+
     It can be used as a ``Flyer`` in a ``bluesky`` plan to have the daq start
     at the beginning of the run and end at the end of the run.
 
-    It has additional knobs for pausing and resuming acquisition during a
-    plan. This can be done by calling `configure` with the ``mode`` parameter:
-
-    - ``on``:      Always take events during the run
-    - ``manual``:  Take events when `calib_cycle` is used
-    - ``auto``:    Take events between ``create`` and ``save`` messages
-
-    Unlike a normal ``bluesky`` ``Flyer``, this has no data to report to the
-    ``RunEngine`` on the ``collect`` call. No data will pass into the python
-    layer from the daq.
+    Unlike normal ``bluesky`` readable devices or flyers, this has no data to
+    report to the ``RunEngine`` on the ``read`` or ``collect`` calls. No data
+    will pass into the python layer from the daq.
 
     Parameters
     ----------
@@ -69,14 +65,13 @@ class Daq(FlyerInterface):
     _state_enum = enum.Enum('PydaqState',
                             'Disconnected Connected Configured Open Running',
                             start=0)
-    _mode_enum = enum.Enum('ScanMode', 'on manual auto', start=0)
     default_config = dict(events=None,
                           duration=None,
                           use_l3t=False,
                           record=False,
-                          controls=None,
-                          mode=_mode_enum.on)
+                          controls=None)
     name = 'daq'
+    parent = None
 
     def __init__(self, platform=None, RE=None):
         if pydaq is None:
@@ -91,8 +86,8 @@ class Daq(FlyerInterface):
         self._desired_config = {}
         self._reset_begin()
         self._host = os.uname()[1]
-        self._is_bluesky = False
         self._RE = RE
+        self._re_cbid = None
         self._config_ts = {}
         self._update_config_ts()
         register_daq(self)
@@ -121,6 +116,18 @@ class Daq(FlyerInterface):
             return self._config.copy()
         else:
             return self.default_config.copy()
+
+    @property
+    def next_config(self):
+        """
+        The next queued configuration.
+
+        This can be different than `config` if we have queued up a
+        configuration to be run on the next begin.
+        """
+        cfg = self.config.copy()
+        cfg.update(self._desired_config)
+        return cfg
 
     @property
     def state(self):
@@ -287,6 +294,41 @@ class Daq(FlyerInterface):
         self.stop()
         self._control.endrun()
 
+    # Reader interface
+    @check_connect
+    def trigger(self):
+        """
+        Begin acquisition. This method blocks until the run begins.
+
+        Returns a status object that will be marked done when the daq has
+        stopped acquiring.
+
+        This will raise a RuntimeError if the daq is configured to run forever.
+
+        Returns
+        -------
+        done_status: ``Status``
+            ``Status`` that will be marked as done when the daq has begun.
+        """
+        cfg = self.next_config
+        if all(cfg[key] is None for key in ('events', 'duration')):
+            raise RuntimeError(('Cannot start daq in scan step, configured '
+                                'to run forever!'))
+        self.begin()
+        return self._get_end_status()
+
+    def read(self):
+        """
+        Return data. There is no data implemented yet.
+        """
+        return {}
+
+    def describe(self):
+        """
+        Explain what read returns. There is nothing  yet.
+        """
+        return {}
+
     # Flyer interface
     @check_connect
     def kickoff(self, events=None, duration=None, use_l3t=None, controls=None):
@@ -398,8 +440,7 @@ class Daq(FlyerInterface):
 
     def collect(self):
         """
-        End the run as part of the ``bluesky`` ``Flyer`` interface, when the
-        flying is done.
+        Collect data as part of the ``bluesky`` ``Flyer`` interface.
 
         As per the ``bluesky`` interface, this is a generator that is expected
         to output partial event documents. However, since we don't have any
@@ -407,9 +448,7 @@ class Daq(FlyerInterface):
         ends.
         """
         logger.debug('Daq.collect()')
-        self.end_run()
-        return
-        yield
+        yield from ()
 
     def describe_collect(self):
         """
@@ -422,7 +461,7 @@ class Daq(FlyerInterface):
 
     @check_connect
     def configure(self, events=None, duration=None, record=None,
-                  use_l3t=False, controls=None, mode=None):
+                  use_l3t=False, controls=None):
         """
         Changes the daq's configuration for the next run.
 
@@ -456,24 +495,13 @@ class Daq(FlyerInterface):
             values each time begin is called. To provide a list, all devices
             must have a ``name`` attribute.
 
-        mode: ``str`` or ``int``, optional
-            This determines our run control during a ``bluesky`` scan with a
-            ``RunEngine`` attached to a Daq object. There are three modes, with
-            the ``on`` mode as the default:
-
-            - ``on``     (0): Take events always during runs
-            - ``manual`` (1): Take events only during the `calib_cycle` plan
-            - ``auto``   (2): Take events between ``create`` and ``save``
-
-            If ``mode`` is omitted, we'll use the previous value for ``mode``.
-
         Returns
         -------
         old, new: ``tuple`` of ``dict``
         """
         logger.debug(('Daq.configure(events=%s, duration=%s, record=%s, '
-                      'use_l3t=%s, controls=%s, mode=%s)'),
-                     events, duration, record, use_l3t, controls, mode)
+                      'use_l3t=%s, controls=%s)'),
+                     events, duration, record, use_l3t, controls)
         state = self.state
         if state not in ('Connected', 'Configured'):
             err = 'Cannot configure from state {}!'.format(state)
@@ -488,18 +516,6 @@ class Daq(FlyerInterface):
         if record is None:
             record = self.config['record']
 
-        if mode is None:
-            mode = self.config['mode']
-        if not isinstance(mode, self._mode_enum):
-            try:
-                mode = getattr(self._mode_enum, mode)
-            except (TypeError, AttributeError):
-                try:
-                    mode = self._mode_enum(mode)
-                except ValueError:
-                    err = '{} is not a valid scan mode!'
-                    raise ValueError(err.format(mode))
-
         config_args = self._config_args(record, use_l3t, controls)
         try:
             logger.debug('Daq.control.configure(%s)',
@@ -509,7 +525,7 @@ class Daq(FlyerInterface):
             # this is different than the arguments that pydaq.Control expects
             self._config = dict(events=events, duration=duration,
                                 record=record, use_l3t=use_l3t,
-                                controls=controls, mode=mode)
+                                controls=controls)
             self._update_config_ts()
             msg = 'Daq configured'
             logger.info(msg)
@@ -608,13 +624,6 @@ class Daq(FlyerInterface):
         logger.debug('Daq._begin_args(%s, %s, %s, %s)',
                      events, duration, use_l3t, controls)
         begin_args = {}
-        if all((self._is_bluesky,
-                not self.config['mode'] == self._mode_enum.on,
-                not self.state == 'Open')):
-            # Open a run without taking events
-            events = 1
-            duration = None
-            use_l3t = False
         if events is None and duration is None:
             events = self.config['events']
             duration = self.config['duration']
@@ -689,15 +698,13 @@ class Daq(FlyerInterface):
                     controls=dict(source='daq_control_vars',
                                   dtype='array',
                                   shape=controls_shape),
-                    always_on=dict(source='daq_mode',
-                                   dtype='number',
-                                   shape=None))
+                    )
 
     def stage(self):
         """
         ``bluesky`` interface for preparing a device for action.
-        There is nothing to be done here, but we overwrite the default stage
-        because we have no sub devices.
+
+        This sets up the daq to end runs on run stop documents.
 
         Returns
         -------
@@ -705,13 +712,20 @@ class Daq(FlyerInterface):
             list of devices staged
         """
         logger.debug('Daq.stage()')
+        if self._re_cbid is None:
+            self._re_cbid = self._RE.subscribe(self._re_manage_runs)
         return [self]
+
+    def _re_manage_runs(self, name, doc):
+        """
+        Callback for the RunEngine to manage run start and stop.
+        """
+        if name == 'stop':
+            self.end_run()
 
     def unstage(self):
         """
         ``bluesky`` interface for undoing the `stage` routine.
-        There is nothing to be done here, but we overwrite the default unstage
-        because we have no sub devices.
 
         Returns
         -------
@@ -719,6 +733,12 @@ class Daq(FlyerInterface):
             list of devices unstaged
         """
         logger.debug('Daq.unstage()')
+        if self._re_cbid is not None:
+            self._RE.unsubscribe(self._re_cbid)
+            self._re_cbid = None
+        # If we're still running, end now
+        if self.state in ('Open', 'Running'):
+            self.end_run()
         return [self]
 
     def pause(self):
@@ -738,32 +758,6 @@ class Daq(FlyerInterface):
         logger.debug('Daq.resume()')
         if self.state == 'Open':
             self.begin()
-
-    def _interpret_message(self, msg):
-        """
-        ``msg_hook`` for the daq to snoop on the ``RunEngine``. This will
-        always start a run at ``open_run`` and end a run at ``close_run``,
-        and in ``auto`` mode this will also `pause` at ``save`` messages and
-        `resume` at ``create`` messages.
-        """
-        logger.debug('Daq._interpret_message(%s)', msg)
-        cmds = ('open_run', 'close_run', 'create', 'save')
-        if msg.command not in cmds:
-            return
-        if msg.command == 'open_run':
-            self._is_bluesky = True
-        elif msg.command == 'close_run':
-            self._is_bluesky = False
-        if self.config['mode'] == self._mode_enum.auto:
-            if msg.command == 'create':
-                # If already runing, pause first to start a fresh begin
-                self.pause()
-                self.resume()
-            elif msg.command == 'save':
-                if self._events or self._duration:
-                    self.wait()
-                else:
-                    self.pause()
 
     @property
     def _events(self):